requirements.txt
# requirements.txt
openai>=1.33.0
anthropic>=0.13.0
mistralai>=0.0.7
cohere>=4.37
google-generativeai>=0.3.1
together>=0.1.5
deepseek>=0.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy
pytest
requests
tqdm
pre-commit
instructor
